refactor some common functions w/Cline
